c files done

+++ srcs/client.c
@@ -4,32 +4,48 @@
 #include <unistd.h>
 #include "minitalk.h"
 
+int	ft_strisdigit(char *str)
+{
+	int i;
+
+	i = 0;
+	if (str[i] == '-')
+		i++;
+	while (str[i])
+	{
+
+		if (!ft_isdigit(str[i]))
+			return (0);
+		i++;
+	}
+	return (1);
+}
+
 void	feedback(int sig)
 {
 	(void)sig;
-	write (1, "Message received by server\n", 28);
+	ft_printf("SUCCESS: String received by server\n");
 	exit (0);
 }
 
-void	client_sighandler(int pid, char *message)
+void	client_sighandler(int pid, char *string)
 {
 	int	bit_count;
 	int	byte_count;
 	char c;
 
 	byte_count = 0;
-	while (message[byte_count] - 1) // WHY minus 1?
+	while (string[byte_count] - 1)
 	{
-		c = message[byte_count];
+		c = string[byte_count];
 		bit_count = 8;
 		while (--bit_count >= 0)
 		{
-			usleep(30); // with this only prints properly
-			if ((c >> bit_count) & 1)
+			if (c && ((c >> bit_count) & 1))
 				kill(pid, SIGUSR1);
 			else
 				kill(pid, SIGUSR2);
-			usleep(30);
+			usleep(60);
 		}
 		byte_count++;
 	}
@@ -38,8 +54,11 @@ void	client_sighandler(int pid, char *message)
 int main (int argc, char **argv)
 {
 	signal(SIGUSR1, feedback);
-	if (argc != 3)
-		printf("invalid");
+	if (argc != 3 || !ft_strisdigit(argv[1]) || !*argv[2])
+	{
+		ft_printf("ERROR: Wrong Format\n");
+		ft_printf("TRY: ./client <PID Number> <String>");
+	}
 	else
 		client_sighandler(ft_atoi(argv[1]), argv[2]);
 }
